mind-explorer-plugin/extern/libworker/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
bson==0.5.10
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
exceptiongroup==1.1.0
fastrlock==0.8.1
fonttools==4.38.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
multidict==6.0.4
nats-py==2.2.0
numpy==1.24.3
packaging==23.0
pandas==1.5.3
pefile==2023.2.7
Pillow==9.5.0
pluggy==1.0.0
pydantic==1.10.5
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
pyzmq==25.0.1
scipy==1.10.0
six==1.16.0
SQLAlchemy==2.0.6
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
win32-setctime==1.1.0
yarl==1.8.2
